+++ verticapy/io/parsers/_utils.py
@@ -42,7 +42,6 @@ def extract_compression(path: str):
     else:
         return "UNCOMPRESSED"
 
-
 def get_first_file(path: str, ext: str):
     directory_name = os.path.dirname(path)
     files = os.listdir(directory_name)
