+++ verticapy/io/parsers/all.py
@@ -22,7 +22,7 @@ import warnings
 from verticapy.utils._decorators import save_verticapy_logs, check_minimum_version
 from verticapy.utils._toolbox import *
 from verticapy.errors import ExtensionError, ParameterError
-from verticapy.io.sql.utils._format import quote_ident, format_schema_table, clean_query
+from verticapy.io.sql._utils._format import quote_ident, format_schema_table, clean_query
 from verticapy.io.parsers._utils import extract_compression, extract_col_dt_from_query
