+++ verticapy/io/sql/read.py
@@ -114,8 +114,9 @@ def to_tablesample(
     tablesample : Object in memory created for rendering purposes.
     """
     import verticapy as vp
-    from verticapy.core.tablesample import tablesample
+    from verticapy.core._utils.tablesample import tablesample
     from verticapy.utils._toolbox import executeSQL, get_final_vertica_type, print_time
+    from verticapy.io.sql._utils._display import print_query
 
     if vp.OPTIONS["sql_on"]:
         print_query(query, title)
@@ -187,7 +188,7 @@ vDataFrame
     from ..flex import isvmap
     from verticapy.io.sql.dtypes import get_data_types
     from verticapy.utils._toolbox import get_category_from_vertica_type
-    from verticapy.io.sql.utils._format import quote_ident
+    from verticapy.io.sql._utils._format import quote_ident
 
     if isinstance(vdf, vDataFrame):
         external = vdf._VERTICAPY_VARIABLES_["external"]
