+++ verticapy/jupyter/extensions/sql_magic.py
@@ -48,14 +48,14 @@ from verticapy import (
     save_verticapy_logs,
 )
 from verticapy.sdk.vertica.dblink import replace_external_queries_in_query
-from verticapy.io.sql.utils._format import replace_vars_in_query, clean_query
+from verticapy.io.sql._utils._format import replace_vars_in_query, clean_query
 
 
 @save_verticapy_logs
 @needs_local_scope
 def sql_magic(line, cell="", local_ns=None):
     from verticapy import vDataFrameSQL
-    from .utils import get_magic_options
+    from ._utils import get_magic_options
 
     # We don't want to display the query/time twice if the options are still on
     # So we save the previous configuration and turn them off.
