+++ verticapy/machine_learning/vertica/delphi.py
@@ -42,7 +42,8 @@ from verticapy.learn.neighbors import *
 from verticapy.learn.svm import *
 from verticapy.plotting._matplotlib import plot_bubble_ml
 from verticapy.learn.vmodel import *
-from verticapy.io.sql.utils._format import schema_relation
+from verticapy.io.sql._utils._format import schema_relation
+from verticapy.machine_learning._utils import reverse_score
 
 # Other Modules
 from tqdm.auto import tqdm
