+++ verticapy/plotting/_highcharts/highchart.py
@@ -533,7 +533,7 @@ def hchart_from_vdf(
         return spider(query=query, options=options, width=width, height=height)
     elif kind in ("pearson", "kendall", "cramer", "biserial", "spearman", "spearmand",):
         data = vdf.corr(method=kind, show=False, columns=x)
-        narrow_data = get_narrow_tablesample(data, use_number_as_category=True)
+        narrow_data = data.narrow(use_number_as_category=True)
         for idx, elem in enumerate(narrow_data[0]):
             try:
                 narrow_data[0][idx][2] = round(elem[2], 2)
