+++ verticapy/plotting/_matplotlib/heatmap.py
@@ -35,7 +35,7 @@ from verticapy.utils._toolbox import (
 from verticapy.core.str_sql import str_sql
 from verticapy.errors import ParameterError
 from verticapy.plotting._colors import gen_colors, gen_cmap
-from verticapy.io.sql.utils._format import quote_ident
+from verticapy.io.sql._utils._format import quote_ident
 
 
 def cmatrix(
