+++ verticapy/stats/math/math.py
@@ -28,7 +28,8 @@ from verticapy.utilities import *
 from verticapy.utils._toolbox import *
 from verticapy.stats.tools import *
 from verticapy.core.str_sql import str_sql
-from verticapy.io.sql.utils._format import format_magic, clean_query
+from verticapy.io.sql._utils._format import format_magic, clean_query
+from verticapy.utils._map import python_to_dtype_category
 
 #
 # Global Variables
@@ -1019,7 +1020,7 @@ str_sql
         raise ParameterError(
             "The number of arguments of the 'case_when' function must be strictly greater than 1."
         )
-    category = get_category_from_python_type(argv[1])
+    category = python_to_dtype_category(argv[1])
     i = 0
     expr = "CASE"
     while i < n:
@@ -1090,7 +1091,7 @@ Returns
 str_sql
     SQL expression.
     """
-    category = get_category_from_python_type(expr)
+    category = python_to_dtype_category(expr)
     expr = [format_magic(expr)]
     for arg in argv:
         expr += [format_magic(arg)]
@@ -1267,7 +1268,7 @@ str_sql
         raise ParameterError(
             "The number of arguments of the 'decode' function must be greater than 3."
         )
-    category = get_category_from_python_type(argv[1])
+    category = python_to_dtype_category(argv[1])
     expr = (
         "DECODE("
         + str(format_magic(expr))
