+++ verticapy/tests/vModel/test_xgb_classifier.py
@@ -32,7 +32,7 @@ from verticapy import vDataFrame, drop, set_option
 from verticapy.connect import current_cursor
 from verticapy.datasets import load_titanic, load_dataset_cl
 from verticapy.learn.ensemble import XGBoostClassifier
-from verticapy.io.sql.utils._format import clean_query
+from verticapy.io.sql._utils._format import clean_query
 
 set_option("print_info", False)
